First pass integration of assessment items with indexeddb.
