+++ contentcuration/contentcuration/frontend/channelEdit/constants.js
@@ -5,3 +5,22 @@ export const RouterNames = {
   CONTENTNODE_DETAILS: 'CONTENTNODE_DETAILS',
   MULTI_CONTENTNODE_DETAILS: 'MULTI_CONTENTNODE_DETAILS',
 };
+
+export const ValidationErrors = {
+  TITLE_REQUIRED: 'TITLE_REQUIRED',
+  LICENCE_REQUIRED: 'LICENCE_REQUIRED',
+  COPYRIGHT_HOLDER_REQUIRED: 'COPYRIGHT_HOLDER_REQUIRED',
+  LICENCE_DESCRIPTION_REQUIRED: 'LICENCE_DESCRIPTION_REQUIRED',
+  MASTERY_MODEL_REQUIRED: 'MASTERY_MODEL_REQUIRED',
+  MASTERY_MODEL_INVALID: 'MASTERY_MODEL_INVALID',
+  QUESTION_REQUIRED: 'QUESTION_REQUIRED',
+  INVALID_NUMBER_OF_CORRECT_ANSWERS: 'INVALID_NUMBER_OF_CORRECT_ANSWERS',
+};
+
+// should correspond to backend types
+export const AssessmentItemTypes = {
+  SINGLE_SELECTION: 'single_selection',
+  MULTIPLE_SELECTION: 'multiple_selection',
+  TRUE_FALSE: 'true_false',
+  INPUT_QUESTION: 'input_question',
+};
