+++ contentcuration/contentcuration/frontend/channelEdit/store.js
@@ -1,4 +1,5 @@
 import template from './vuex/template';
+import assessmentItem from './vuex/assessmentItem';
 import contentNode from './vuex/contentNode';
 import currentChannel from './vuex/currentChannel';
 import storeFactory from 'shared/vuex/baseStore';
@@ -6,6 +7,7 @@ import storeFactory from 'shared/vuex/baseStore';
 const store = storeFactory({
   modules: {
     template,
+    assessmentItem,
     contentNode,
     currentChannel,
   },
