+++ contentcuration/contentcuration/frontend/channelEdit/vuex/assessmentItem/__tests__/utils.spec.js
@@ -1,0 +1,397 @@
+import { AssessmentItemTypes, ValidationErrors } from '../../../constants';
+import {
+  sanitizeAssessmentItemAnswers,
+  sanitizeAssessmentItemHints,
+  sanitizeAssessmentItem,
+  validateAssessmentItem,
+  sanitizeAssessmentItems,
+} from '../utils';
+
+describe('utils', () => {
+
+  describe('sanitizeAssessmentItemAnswers', () => {
+    it('trims answers', () => {
+      const answers = [
+        { answer: '', order: 1, correct: true },
+        { answer: ' 3 ', order: 2, correct: false },
+        { answer: '  ', order: 3, correct: true },
+      ];
+
+      expect(sanitizeAssessmentItemAnswers(answers)).toEqual([
+        { answer: '', order: 1, correct: true },
+        { answer: '3', order: 2, correct: false },
+        { answer: '', order: 3, correct: true },
+      ]);
+    });
+
+    it('removes all empty answers and reorders remaining answers if removeEmpty true', () => {
+      const answers = [
+        { answer: '', order: 1, correct: true },
+        { answer: ' 3 ', order: 2, correct: false },
+        { answer: '  ', order: 3, correct: true },
+      ];
+
+      expect(sanitizeAssessmentItemAnswers(answers, true)).toEqual([
+        { answer: '3', order: 1, correct: false },
+      ]);
+    });
+  });
+
+  describe('sanitizeAssessmentItemHints', () => {
+    it('trims hints', () => {
+      const hints = [
+        { hint: '', order: 1 },
+        { hint: ' Hint 1 ', order: 2 },
+        { hint: '  ', order: 3 },
+      ];
+
+      expect(sanitizeAssessmentItemHints(hints)).toEqual([
+        { hint: '', order: 1 },
+        { hint: 'Hint 1', order: 2 },
+        { hint: '', order: 3 },
+      ]);
+    });
+
+    it('removes all empty hints and reorders remaining hints if removeEmpty true', () => {
+      const hints = [
+        { hint: '', order: 1 },
+        { hint: ' Hint 1 ', order: 2 },
+        { hint: '  ', order: 3 },
+      ];
+
+      expect(sanitizeAssessmentItemHints(hints, true)).toEqual([{ hint: 'Hint 1', order: 1 }]);
+    });
+  });
+
+  describe('sanitizeAssessmentItem', () => {
+    it('trims question, hints and answers', () => {
+      const assessmentItem = {
+        order: 1,
+        question: ' Question text ',
+        answers: [
+          { answer: ' Answer 1', order: 1, correct: false },
+          { answer: '', order: 2, correct: true },
+          { answer: 'Answer 3    ', order: 3, correct: true },
+        ],
+        hints: [{ hint: ' ', order: 1 }, { hint: '', order: 2 }, { hint: ' Hint 3', order: 3 }],
+      };
+
+      expect(sanitizeAssessmentItem(assessmentItem)).toEqual({
+        order: 1,
+        question: 'Question text',
+        answers: [
+          { answer: 'Answer 1', order: 1, correct: false },
+          { answer: '', order: 2, correct: true },
+          { answer: 'Answer 3', order: 3, correct: true },
+        ],
+        hints: [{ hint: '', order: 1 }, { hint: '', order: 2 }, { hint: 'Hint 3', order: 3 }],
+      });
+    });
+
+    it('removes all empty hints and answers if removeEmpty true', () => {
+      const assessmentItem = {
+        order: 1,
+        question: ' Question text ',
+        answers: [
+          { answer: ' Answer 1', order: 1, correct: false },
+          { answer: '', order: 2, correct: true },
+          { answer: 'Answer 3    ', order: 3, correct: true },
+        ],
+        hints: [{ hint: ' ', order: 1 }, { hint: '', order: 2 }, { hint: ' Hint 3', order: 3 }],
+      };
+
+      expect(sanitizeAssessmentItem(assessmentItem, true)).toEqual({
+        order: 1,
+        question: 'Question text',
+        answers: [
+          { answer: 'Answer 1', order: 1, correct: false },
+          { answer: 'Answer 3', order: 2, correct: true },
+        ],
+        hints: [{ hint: 'Hint 3', order: 1 }],
+      });
+    });
+  });
+
+  describe('validateAssessmentItem', () => {
+    describe('when question text is missing', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: '',
+          answers: [{ answer: 'Answer', correct: true, order: 1 }],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.QUESTION_REQUIRED,
+        ]);
+      });
+    });
+
+    describe('for single selection with no answers', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.SINGLE_SELECTION,
+          answers: [],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for single selection with no correct answer', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.SINGLE_SELECTION,
+          answers: [{ answer: 'Answer', correct: false, order: 1 }],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for single selection with more correct answers', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.SINGLE_SELECTION,
+          answers: [
+            { answer: 'Answer 1', correct: true, order: 1 },
+            { answer: 'Answer 2', correct: true, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for single selection with one correct answer', () => {
+      it('returns positive validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.SINGLE_SELECTION,
+          answers: [
+            { answer: 'Answer 1', correct: false, order: 1 },
+            { answer: 'Answer 2', correct: true, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([]);
+      });
+    });
+
+    describe('for multiple selection with no answers', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.MULTIPLE_SELECTION,
+          answers: [],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for multiple selection with no correct answer', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.MULTIPLE_SELECTION,
+          answers: [
+            { answer: 'Answer 1', correct: false, order: 1 },
+            { answer: 'Answer 2', correct: false, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for multiple selection with at least one correct answer', () => {
+      it('returns positive validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.MULTIPLE_SELECTION,
+          answers: [
+            { answer: 'Answer 1', correct: true, order: 1 },
+            { answer: 'Answer 2', correct: false, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([]);
+      });
+    });
+
+    describe('for input question with no answers', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.INPUT_QUESTION,
+          answers: [],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for input question with no correct answer', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.INPUT_QUESTION,
+          answers: [
+            { answer: 'Answer 1', correct: false, order: 1 },
+            { answer: 'Answer 2', correct: false, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for input question with at least one correct answer', () => {
+      it('returns positive validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.INPUT_QUESTION,
+          answers: [
+            { answer: 'Answer 1', correct: true, order: 1 },
+            { answer: 'Answer 2', correct: true, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([]);
+      });
+    });
+
+    describe('for true/false with no answers', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.TRUE_FALSE,
+          answers: [],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for true/false with no correct answer', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.TRUE_FALSE,
+          answers: [
+            { answer: 'True', correct: false, order: 1 },
+            { answer: 'False', correct: false, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for true/false with more correct answers', () => {
+      it('returns negative validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.TRUE_FALSE,
+          answers: [
+            { answer: 'True', correct: true, order: 1 },
+            { answer: 'False', correct: true, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([
+          ValidationErrors.INVALID_NUMBER_OF_CORRECT_ANSWERS,
+        ]);
+      });
+    });
+
+    describe('for true/false with one correct answer', () => {
+      it('returns positive validation results', () => {
+        const assessmentItem = {
+          question: 'Question',
+          type: AssessmentItemTypes.TRUE_FALSE,
+          answers: [
+            { answer: 'True', correct: false, order: 1 },
+            { answer: 'False', correct: true, order: 2 },
+          ],
+        };
+
+        expect(validateAssessmentItem(assessmentItem)).toEqual([]);
+      });
+    });
+  });
+
+  describe('sanitizeAssessmentItems', () => {
+    it('sanitizes questions/answers/hints texts and removes empty questions/answers/hints', () => {
+      const assessmentItems = [
+        {
+          order: 0,
+          question: ' Question 1 text ',
+          answers: [
+            { answer: ' Answer 1', order: 1, correct: false },
+            { answer: '', order: 2, correct: true },
+            { answer: 'Answer 3 ', order: 3, correct: true },
+          ],
+          hints: [{ hint: ' ', order: 1 }, { hint: '', order: 2 }, { hint: ' Hint 3', order: 3 }],
+        },
+        {
+          order: 1,
+          question: ' ',
+          answers: [{ answer: '', order: 1, correct: true }],
+          hints: [{ hint: ' ', order: 1 }, { hint: '', order: 2 }],
+        },
+        {
+          order: 2,
+          question: '',
+          answers: [],
+          hints: [],
+        },
+        {
+          order: 3,
+          question: ' Question 4 text ',
+          answers: [{ answer: '', order: 2, correct: true }],
+          hints: [],
+        },
+      ];
+
+      expect(sanitizeAssessmentItems(assessmentItems)).toEqual([
+        {
+          order: 0,
+          question: 'Question 1 text',
+          answers: [
+            { answer: 'Answer 1', order: 1, correct: false },
+            { answer: 'Answer 3', order: 2, correct: true },
+          ],
+          hints: [{ hint: 'Hint 3', order: 1 }],
+        },
+        {
+          order: 1,
+          question: 'Question 4 text',
+          answers: [],
+          hints: [],
+        },
+      ]);
+    });
+  });
+});
