+++ contentcuration/contentcuration/frontend/channelEdit/vuex/contentNode/mutations.js
@@ -1,27 +1,18 @@
 import Vue from 'vue';
-
-function mergeNode(nodesMap, node) {
-  return {
-    ...nodesMap,
-    [node.id]: {
-      ...nodesMap[node.id],
-      ...node,
-    },
-  };
-}
+import { mergeMapItem } from 'shared/vuex/utils';
 
 export function ADD_CONTENTNODE(state, contentNode) {
-  state.contentNodesMap = mergeNode(state.contentNodesMap, contentNode);
+  state.contentNodesMap = mergeMapItem(state.contentNodesMap, contentNode);
 }
 
 export function ADD_CONTENTNODES(state, contentNodes = []) {
   state.contentNodesMap = contentNodes.reduce((contentNodesMap, contentNode) => {
-    return mergeNode(contentNodesMap, contentNode);
+    return mergeMapItem(contentNodesMap, contentNode);
   }, state.contentNodesMap);
 }
 
-export function REMOVE_CONTENTNODE(state, contentNodeId) {
-  Vue.delete(state.contentNodesMap, contentNodeId);
+export function REMOVE_CONTENTNODE(state, contentNode) {
+  Vue.delete(state.contentNodesMap, contentNode.id);
 }
 
 export function UPDATE_CONTENTNODE(state, { id, ...payload } = {}) {
@@ -63,17 +54,17 @@ export function TOGGLE_EXPANSION(state, id) {
 }
 
 export function ADD_TREENODE(state, treeNode) {
-  state.treeNodesMap = mergeNode(state.treeNodesMap, treeNode);
+  state.treeNodesMap = mergeMapItem(state.treeNodesMap, treeNode);
 }
 
 export function ADD_TREENODES(state, treeNodes = []) {
   state.treeNodesMap = treeNodes.reduce((treeNodesMap, treeNode) => {
-    return mergeNode(treeNodesMap, treeNode);
+    return mergeMapItem(treeNodesMap, treeNode);
   }, state.treeNodesMap);
 }
 
-export function REMOVE_TREENODE(state, treeNodeId) {
-  Vue.delete(state.treeNodesMap, treeNodeId);
+export function REMOVE_TREENODE(state, treeNode) {
+  Vue.delete(state.treeNodesMap, treeNode.id);
 }
 
 export function UPDATE_TREENODE(state, { id, ...payload } = {}) {
