+++ contentcuration/contentcuration/frontend/shared/vuex/utils.js
@@ -1,0 +1,9 @@
+export function mergeMapItem(itemsMap, item, idField = 'id') {
+  return {
+    ...itemsMap,
+    [item[idField]]: {
+      ...itemsMap[item[idField]],
+      ...item,
+    },
+  };
+}
