+++ contentcuration/contentcuration/static/js/edit_channel/uploader/views/__tests__/AssessmentView.spec.js
@@ -1,18 +1,13 @@
-import { mount, createLocalVue } from '@vue/test-utils';
-import Vuex, { Store } from 'vuex';
+import { mount } from '@vue/test-utils';
 
 import AssessmentView from '../AssessmentView';
 import { AssessmentItemTypes, ValidationErrors } from '../../constants';
-
-// TODO @MisRob: Consistent imports
-const editModalGetters = require('../../vuex/getters');
-const editModalMutations = require('../../vuex/mutations');
+import assessmentItem from 'frontend/channelEdit/vuex/assessmentItem';
+import contentNode from 'frontend/channelEdit/vuex/contentNode';
+import storeFactory from 'shared/vuex/baseStore';
 
 jest.mock('../../components//MarkdownEditor/MarkdownEditor/MarkdownEditor.vue');
 jest.mock('../../components//MarkdownEditor/MarkdownViewer/MarkdownViewer.vue');
-
-const localVue = createLocalVue();
-localVue.use(Vuex);
 
 const SELECTED_NODE_IDX = 1;
 
@@ -93,20 +88,15 @@ const EDIT_MODAL_STATE = {
   },
 };
 
-const initWrapper = state => {
-  const store = new Store({
+const initWrapper = () => {
+  const store = storeFactory({
     modules: {
-      edit_modal: {
-        namespaced: true,
-        state,
-        getters: editModalGetters,
-        mutations: editModalMutations,
-      },
-    },
+      assessmentItem,
+      contentNode,
+    }
   });
 
   return mount(AssessmentView, {
-    localVue,
     store,
   });
 };
